requirements.txt
streamlit
supabase
google-generativeai
moviepy